backend/requirements.txt
streamlit==1.27.2
pandas==2.1.1
openai==0.28.1
plotly==5.17.0
python-dotenv==1.0.0